requirements.txt
chainlit
openai
python-dotenv
playwright
httpx
PyJWT
cryptography 